Remove dev options, change chat font size, remove console logs, change default visibility, group components by folders
