+++ src/components/icons/ClipboardIcon.vue
