+++ src/components/icons/EmoteIcon.vue
