+++ src/components/icons/LeftArrowIcon.vue
