+++ src/components/icons/MinusIcon.vue
