+++ src/components/icons/SettingsIcon.vue
