+++ src/modules/socket.ts
@@ -141,13 +141,10 @@ export class ClientSocketHandler {
         );
 
         this.socket.on(SocketEventType.PERMISSIONS, (message: UserPermissions[]) => {
-            this.userPermissionsStore.usersPermissions = message;
-            console.log("received updated permissions", message);
-            // check if current user is present on the list of permissions
+            this.userPermissionsStore.usersPermissions = message;// check if current user is present on the list of permissions
             // if not - he is kicked from the room
-            console.log(message[this.socket.id]);
             if(!message[this.socket.id]) {
-                console.log(console.log("You have been kicked from the room!"));
+                console.log("You have been kicked from the room!");
                 this.socket.disconnect();
             }
         });
@@ -323,14 +320,13 @@ export class ClientSocketHandler {
         this.checkForErrors();
         return new Promise((resolve, reject) => {
             if(this.socket.id === username) {
-                console.log("u cant kick yourself");
                 reject("You can't kick yourself!");
             }
             this.socket.emit(
                 SocketEventType.KICK_USER,
                 username,
                 (response: any) => {
-                    if (response === "ROOM_NOT_FOUND" || 
+                    if (response === "ROOM_NOT_FOUND" ||
                         response === "OPERATION_NOT_ALLOWED") {
                         reject(response);
                     }
